geolocation/test_launcher.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Power API</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body/>

  <script type="application/javascript">

  // It is a requirement that Bug: 797513 gets landed on m/c, before this test will pass

  SimpleTest.waitForExplicitFinish();
  var domain = "http://example.org/"; 
  var url = window.location.href.replace(/:\/\/[^\/]*\//, "://example.org/");
  url = url.substring(0, url.lastIndexOf('/')) + "/allowCurrent.html"; 
  
  SpecialPowers.addPermission("settings", true, domain);
  SpecialPowers.addPermission("geolocation", true, domain);
  SpecialPowers.addPermission("browser", true, domain);


  var iframe = document.createElement("iframe");
  iframe.mozApp = domain + "manifest.webapp";
  iframe.src = url;
  iframe.id = "ifrm";
  iframe.onload = loaded();
  iframe.mozbrowser = true;
  
  document.body.appendChild(iframe);

  function loaded() {

    window.addEventListener("message", function onMessage(event) {
       if (event.data == "finished") {
         window.removeEventListener("message", onMessage, false);
         SimpleTest.finish();
       } else {
         dump(JSON.stringify(event.data));
         ok(event.data.result, event.data.msg);
       }
           
     }, false);

  }
  </script>
</head>

</html>
